+++ src/main/java/pl/kj/bachelors/teams/infrastructure/service/invitation/InvitationManagementService.java
@@ -1,14 +1,16 @@
 package pl.kj.bachelors.teams.infrastructure.service.invitation;
 
-import org.apache.commons.lang3.NotImplementedException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import pl.kj.bachelors.teams.domain.config.TeamInvitationConfig;
+import pl.kj.bachelors.teams.domain.exception.AccessDeniedException;
 import pl.kj.bachelors.teams.domain.exception.ResourceNotFoundException;
 import pl.kj.bachelors.teams.domain.model.entity.Team;
 import pl.kj.bachelors.teams.domain.model.entity.TeamInvitation;
+import pl.kj.bachelors.teams.domain.model.extension.action.TeamInvitationActions;
 import pl.kj.bachelors.teams.domain.service.invitation.InvitationManager;
+import pl.kj.bachelors.teams.domain.service.security.EntityAccessControlService;
 import pl.kj.bachelors.teams.infrastructure.repository.TeamInvitationRepository;
 import pl.kj.bachelors.teams.infrastructure.repository.TeamRepository;
 import pl.kj.bachelors.teams.infrastructure.service.TokenGenerationService;
@@ -22,34 +24,42 @@ public class InvitationManagementService implements InvitationManager {
     private final TeamInvitationRepository invitationRepository;
     private final TeamInvitationConfig config;
     private final TokenGenerationService tokenService;
+    private final EntityAccessControlService<Team> accessControl;
 
     @Autowired
     public InvitationManagementService(
             TeamRepository teamRepository,
             TeamInvitationRepository invitationRepository,
             TeamInvitationConfig config,
-            TokenGenerationService tokenService) {
+            TokenGenerationService tokenService,
+            EntityAccessControlService<Team> accessControl
+    ) {
         this.teamRepository = teamRepository;
         this.invitationRepository = invitationRepository;
         this.config = config;
         this.tokenService = tokenService;
+        this.accessControl = accessControl;
     }
 
     @Override
-    @Transactional(rollbackFor = ResourceNotFoundException.class)
-    public TeamInvitation open(int teamId) throws ResourceNotFoundException, ExecutionException, InterruptedException {
+    @Transactional(rollbackFor = {ResourceNotFoundException.class, AccessDeniedException.class})
+    public TeamInvitation open(int teamId)
+            throws ResourceNotFoundException,
+            ExecutionException,
+            InterruptedException,
+            AccessDeniedException {
         Team team = this.teamRepository.findById(teamId).orElseThrow(ResourceNotFoundException::new);
-
+        this.accessControl.ensureThatUserHasAccess(team, TeamInvitationActions.CREATE);
         return this.createInvitation(team);
     }
 
     @Override
-    @Transactional(rollbackFor = ResourceNotFoundException.class)
-    public void close(String code) throws ResourceNotFoundException {
+    @Transactional(rollbackFor = {ResourceNotFoundException.class, AccessDeniedException.class})
+    public void close(String code) throws ResourceNotFoundException, AccessDeniedException {
         TeamInvitation invitation = this.invitationRepository
                 .findFirstByCode(code)
                 .orElseThrow(ResourceNotFoundException::new);
-
+        this.accessControl.ensureThatUserHasAccess(invitation.getTeam(), TeamInvitationActions.CLOSE);
         invitationRepository.delete(invitation);
     }
